+++ web/src/nav/GlobalNavbar.tsx
@@ -85,7 +85,10 @@ export class GlobalNavbar extends React.PureComponent<Props, State> {
                 ? '/.assets/img/sourcegraph-light-head-logo.svg'
                 : '/.assets/img/sourcegraph-head-logo.svg'
         } else {
-            logoSrc = '/.assets/img/sourcegraph-mark.svg'
+            logoSrc =
+                window.context.branding && window.context.branding.favicon
+                    ? window.context.branding.favicon
+                    : '/.assets/img/sourcegraph-mark.svg'
         }
 
         const logo = (